rename module to 'rdf'
