+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/Info.java
@@ -1,7 +1,7 @@
 package org.molgenis.emx2.beaconv2.endpoints;
 
-import static org.molgenis.emx2.semantics.RDFService.extractHost;
-import static org.molgenis.emx2.semantics.rdf.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.rdf.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.rdf.RDFService.extractHost;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import java.net.URI;
