+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/biosamples/BiosamplesResponse.java
@@ -2,7 +2,7 @@ package org.molgenis.emx2.beaconv2.endpoints.biosamples;
 
 import static org.molgenis.emx2.FilterBean.f;
 import static org.molgenis.emx2.Operator.EQUALS;
-import static org.molgenis.emx2.semantics.QueryHelper.*;
+import static org.molgenis.emx2.rdf.QueryHelper.*;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnore;
