+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/biosamples/BiosamplesResultSetsItem.java
@@ -1,7 +1,7 @@
 package org.molgenis.emx2.beaconv2.endpoints.biosamples;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import org.molgenis.emx2.semantics.OntologyTerm;
+import org.molgenis.emx2.rdf.OntologyTerm;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class BiosamplesResultSetsItem {
