+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/datasets/DatasetsResultSetsItem.java
@@ -1,7 +1,7 @@
 package org.molgenis.emx2.beaconv2.endpoints.datasets;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import org.molgenis.emx2.semantics.OntologyTerm;
+import org.molgenis.emx2.rdf.OntologyTerm;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class DatasetsResultSetsItem {
