+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/datasets/Queries.java
@@ -1,8 +1,8 @@
 package org.molgenis.emx2.beaconv2.endpoints.datasets;
 
-import static org.molgenis.emx2.semantics.RDFService.extractHost;
-import static org.molgenis.emx2.semantics.rdf.IRIParsingEncoding.encodedIRI;
-import static org.molgenis.emx2.semantics.rdf.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.rdf.IRIParsingEncoding.encodedIRI;
+import static org.molgenis.emx2.rdf.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.rdf.RDFService.extractHost;
 
 import graphql.ExecutionResult;
 import graphql.GraphQL;
