+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/individuals/IndividualsResultSetsItem.java
@@ -3,7 +3,7 @@ package org.molgenis.emx2.beaconv2.endpoints.individuals;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import org.molgenis.emx2.beaconv2.common.AgeAndAgeGroup;
 import org.molgenis.emx2.beaconv2.endpoints.genomicvariants.CaseLevelData;
-import org.molgenis.emx2.semantics.OntologyTerm;
+import org.molgenis.emx2.rdf.OntologyTerm;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class IndividualsResultSetsItem {
