+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/individuals/Measures.java
@@ -1,12 +1,12 @@
 package org.molgenis.emx2.beaconv2.endpoints.individuals;
 
-import static org.molgenis.emx2.semantics.QueryHelper.mapToOntologyTerm;
+import static org.molgenis.emx2.rdf.QueryHelper.mapToOntologyTerm;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import java.util.List;
 import java.util.Map;
 import org.molgenis.emx2.beaconv2.common.ISO8601duration;
-import org.molgenis.emx2.semantics.OntologyTerm;
+import org.molgenis.emx2.rdf.OntologyTerm;
 import org.molgenis.emx2.utils.TypeUtils;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
