+++ backend/molgenis-emx2-fairdatapoint/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     implementation project(':backend:molgenis-emx2')
-    implementation project(':backend:molgenis-emx2-semantics')
+    implementation project(':backend:molgenis-emx2-rdf')
     implementation project(':backend:molgenis-emx2-beacon-v2')
     implementation project(':backend:molgenis-emx2-graphql')
     implementation 'org.eclipse.rdf4j:rdf4j-rio-api:4.3.7'
