+++ backend/molgenis-emx2-graphgenome/src/main/java/org/molgenis/emx2/graphgenome/Semantics.java
@@ -134,7 +134,7 @@ public class Semantics {
             builder.add(
                 clinIntNode,
                 "http://purl.obolibrary.org/obo/NCIT_C25372",
-                iri(ci.getCategory().URI));
+                iri(ci.getCategory().uri));
           }
           if (ci.getEffect() != null) {
             builder.add(
@@ -142,7 +142,7 @@ public class Semantics {
                 "http://purl.obolibrary.org/obo/NCIT_C15607",
                 ci.getEffect().getLabel());
             builder.add(
-                clinIntNode, "http://purl.obolibrary.org/obo/NCIT_C15607", iri(ci.getEffect().URI));
+                clinIntNode, "http://purl.obolibrary.org/obo/NCIT_C15607", iri(ci.getEffect().uri));
           }
           if (ci.getClinicalRelevance() != null) {
             builder.add(
@@ -152,7 +152,7 @@ public class Semantics {
             builder.add(
                 clinIntNode,
                 "http://purl.obolibrary.org/obo/HP_0045088",
-                iri(ci.getClinicalRelevance().URI));
+                iri(ci.getClinicalRelevance().uri));
           }
           if (ci.getConditionId() != null) {
             builder.add(clinIntNode, "http://edamontology.org/data_3667", ci.getConditionId());
