+++ backend/molgenis-emx2-rdf/build.gradle
