+++ backend/molgenis-emx2-rdf/src/main/java/org/molgenis/emx2/rdf/ValueToRDF.java
@@ -1,12 +1,10 @@
-package org.molgenis.emx2.semantics.rdf;
+package org.molgenis.emx2.rdf;
 
 import static org.eclipse.rdf4j.model.util.Values.iri;
 import static org.eclipse.rdf4j.model.util.Values.literal;
 import static org.molgenis.emx2.FilterBean.and;
 import static org.molgenis.emx2.FilterBean.f;
 import static org.molgenis.emx2.Operator.EQUALS;
-import static org.molgenis.emx2.semantics.rdf.ColumnTypeToXSDDataType.columnTypeToXSD;
-import static org.molgenis.emx2.semantics.rdf.IRIParsingEncoding.encodedIRI;
 
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
@@ -29,6 +27,48 @@ public class ValueToRDF {
     // hide constructor, not meant to be instantiated
   }
 
+  public static CoreDatatype.XSD columnTypeToXSD(ColumnType columnType) {
+    switch (columnType) {
+      case BOOL, BOOL_ARRAY:
+        return CoreDatatype.XSD.BOOLEAN;
+
+      case DATE, DATE_ARRAY:
+        return CoreDatatype.XSD.DATE;
+
+      case DATETIME, DATETIME_ARRAY:
+        return CoreDatatype.XSD.DATETIME;
+
+      case DECIMAL, DECIMAL_ARRAY:
+        return CoreDatatype.XSD.DECIMAL;
+
+      case EMAIL,
+          EMAIL_ARRAY,
+          HEADING,
+          JSONB,
+          JSONB_ARRAY,
+          STRING,
+          STRING_ARRAY,
+          TEXT,
+          TEXT_ARRAY,
+          UUID,
+          UUID_ARRAY,
+          AUTO_ID:
+        return CoreDatatype.XSD.STRING;
+
+      case FILE, HYPERLINK, HYPERLINK_ARRAY, ONTOLOGY, ONTOLOGY_ARRAY, REF, REF_ARRAY, REFBACK:
+        return CoreDatatype.XSD.ANYURI;
+
+      case INT, INT_ARRAY:
+        return CoreDatatype.XSD.INT;
+
+      case LONG, LONG_ARRAY:
+        return CoreDatatype.XSD.LONG;
+
+      default:
+        throw new MolgenisException("ColumnType not mapped: " + columnType);
+    }
+  }
+
   public static void describeValues(
       ModelBuilder builder, Table table, String rowId, String schemaContext) {
     Map<String, Column> columnMap = new HashMap<>();
@@ -36,6 +76,38 @@ public class ValueToRDF {
       columnMap.put(c.getName(), c);
     }
     String tableContext = schemaContext + "/" + table.getName();
+    for (Row row : getRows(table, rowId)) {
+      IRI rowContext =
+          getIriValuesBasedOnPkey(schemaContext, table.getMetadata(), row, "")
+              .get(0); // note the prefix
+      builder.add(rowContext, RDF.TYPE, IRIParsingEncoding.encodedIRI(tableContext));
+      // SIO:001187 = database row
+      builder.add(rowContext, RDF.TYPE, iri("http://semanticscience.org/resource/SIO_001187"));
+      if (table.getMetadata().getTableType() == TableType.ONTOLOGIES) {
+        // NCIT:C95637 = Coded Value Data Type
+        builder.add(rowContext, RDF.TYPE, iri("http://purl.obolibrary.org/obo/NCIT_C95637"));
+        if (row.getString("ontologyTermURI") != null) {
+          builder.add(rowContext, RDFS.ISDEFINEDBY, iri(row.getString("ontologyTermURI")));
+        }
+      } else {
+        builder.add(rowContext, RDF.TYPE, iri("http://purl.org/linked-data/cube#Observation"));
+      }
+      builder.add(
+          rowContext,
+          iri("http://purl.org/linked-data/cube#dataSet"),
+          IRIParsingEncoding.encodedIRI(tableContext));
+
+      for (Column column : table.getMetadata().getColumns()) {
+        IRI columnContext =
+            IRIParsingEncoding.encodedIRI(tableContext + "/column/" + column.getName());
+        for (Value value : formatValue(row, column, schemaContext)) {
+          builder.add(rowContext, columnContext, value);
+        }
+      }
+    }
+  }
+
+  private static List<Row> getRows(Table table, String rowId) {
     Query query = table.query();
     if (rowId != null) {
       if (table.getMetadata().getPrimaryKeyFields().size() > 1) {
@@ -44,48 +116,13 @@ public class ValueToRDF {
         query.where(f(table.getMetadata().getPrimaryKeyColumns().get(0).getName(), EQUALS, rowId));
       }
     }
-
-    List<Row> rows = query.retrieveRows(); // we use the default select
-    if (rows.isEmpty()) {
-      return;
-    } else {
-
-      for (Row row : rows) {
-        IRI rowContext =
-            getIriValuesBasedOnPkey(schemaContext, table.getMetadata(), row, "")
-                .get(0); // note the prefix
-        builder.add(rowContext, RDF.TYPE, encodedIRI(tableContext));
-        // SIO:001187 = database row
-        builder.add(rowContext, RDF.TYPE, iri("http://semanticscience.org/resource/SIO_001187"));
-        if (table.getMetadata().getTableType() == TableType.ONTOLOGIES) {
-          // NCIT:C95637 = Coded Value Data Type
-          builder.add(rowContext, RDF.TYPE, iri("http://purl.obolibrary.org/obo/NCIT_C95637"));
-          if (row.getString("ontologyTermURI") != null) {
-            builder.add(rowContext, RDFS.ISDEFINEDBY, iri(row.getString("ontologyTermURI")));
-          }
-        } else {
-          builder.add(rowContext, RDF.TYPE, iri("http://purl.org/linked-data/cube#Observation"));
-        }
-        builder.add(
-            rowContext, iri("http://purl.org/linked-data/cube#dataSet"), encodedIRI(tableContext));
-
-        for (Column column : table.getMetadata().getColumns()) {
-          if (row.getString(column.getName()) != null) { // empty lists we don't want
-            IRI columnContext = encodedIRI(tableContext + "/column/" + column.getName());
-            for (Value value : formatValue(row, column, schemaContext)) {
-              if (value != null) {
-                builder.add(rowContext, columnContext, value);
-              }
-            }
-          }
-        }
-      }
-    }
+    return query.retrieveRows(); // we use the default select
   }
 
   private static Filter decodeRowIdToFilter(String rowId) {
     try {
-      List<NameValuePair> params = URLEncodedUtils.parse(new URI(rowId), StandardCharsets.UTF_8);
+      List<NameValuePair> params =
+          URLEncodedUtils.parse(new URI("?" + rowId), StandardCharsets.UTF_8);
       List<Filter> filters = new ArrayList<>();
       params.forEach(param -> filters.add(f(param.getName(), EQUALS, param.getValue())));
       return and(filters);
@@ -106,7 +143,9 @@ public class ValueToRDF {
     if (tableMetadata.getPrimaryKeyFields().size() == 1) {
       return Arrays.stream(keys)
           .map(
-              value -> encodedIRI(schemaContext + "/" + tableMetadata.getTableName() + "/" + value))
+              value ->
+                  IRIParsingEncoding.encodedIRI(
+                      schemaContext + "/" + tableMetadata.getTableName() + "/" + value))
           .toList();
     }
     // composite keys get pattern of part1=a&part2=b
@@ -115,7 +154,7 @@ public class ValueToRDF {
       tableMetadata.getPrimaryKeyFields().stream()
           .forEach(
               field -> {
-                String fieldName = prefix + field.getName(); // todo should not be needed
+                String fieldName = prefix + field.getName();
                 String[] keyParts = row.getStringArray(fieldName);
                 for (int i = 0; i < keyParts.length; i++) {
                   if (keyValuePairList.size() <= i) keyValuePairList.add(new ArrayList<>());
@@ -125,7 +164,7 @@ public class ValueToRDF {
       return keyValuePairList.stream()
           .map(
               valuePairList ->
-                  encodedIRI(
+                  IRIParsingEncoding.encodedIRI(
                       schemaContext
                           + "/"
                           + tableMetadata.getTableName()
@@ -137,16 +176,15 @@ public class ValueToRDF {
 
   public static List<Value> formatValue(Row row, Column column, String schemaContext) {
     List<Value> values = new ArrayList<>();
-
     ColumnType columnType = column.getColumnType();
     if (columnType.isReference()) {
       values.addAll(
           getIriValuesBasedOnPkey(
-              schemaContext, column.getRefTable(), row, column.getName() + "-"));
+              schemaContext, column.getRefTable(), row, column.getName() + "."));
     } else if (columnType.equals(ColumnType.FILE)) {
       if (row.getString(column.getName() + "_id") != null) {
         values.add(
-            encodedIRI(
+            IRIParsingEncoding.encodedIRI(
                 schemaContext
                     + "/api/file/"
                     + column.getTableName()
@@ -162,6 +200,9 @@ public class ValueToRDF {
 
   public static List<? extends Value> getLiteralValues(Row row, Column column) {
     CoreDatatype.XSD xsdType = columnTypeToXSD(column.getColumnType());
+    if (row.getString(column.getName()) == null) {
+      return List.of();
+    }
     switch (xsdType) {
       case BOOLEAN:
         return Arrays.stream(row.getBooleanArray(column.getName())).map(Values::literal).toList();
@@ -179,7 +220,7 @@ public class ValueToRDF {
         return Arrays.stream(row.getStringArray(column.getName())).map(Values::literal).toList();
       case ANYURI:
         return Arrays.stream(row.getStringArray(column.getName()))
-            .map(value -> encodedIRI(value))
+            .map(value -> IRIParsingEncoding.encodedIRI(value))
             .toList();
       case INT:
         return Arrays.stream(row.getIntegerArray(column.getName())).map(Values::literal).toList();
