+++ backend/molgenis-emx2-rdf/src/test/java/org/molgenis/emx2/rdf/RDFTest.java
@@ -1,7 +1,7 @@
-package org.molgenis.emx2.semantics.rdf;
+package org.molgenis.emx2.rdf;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.molgenis.emx2.semantics.rdf.StringsForRDFTest.*;
+import static org.molgenis.emx2.rdf.StringsForRDFTest.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
@@ -11,7 +11,6 @@ import org.molgenis.emx2.Database;
 import org.molgenis.emx2.Schema;
 import org.molgenis.emx2.Table;
 import org.molgenis.emx2.datamodels.PetStoreLoader;
-import org.molgenis.emx2.semantics.RDFService;
 import org.molgenis.emx2.sql.TestDatabaseFactory;
 
 public class RDFTest {
