+++ backend/molgenis-emx2-rdf/src/test/java/org/molgenis/emx2/rdf/RDFWithCompositeKeysTest.java
@@ -1,10 +1,10 @@
-package org.molgenis.emx2.semantics.rdf;
+package org.molgenis.emx2.rdf;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.molgenis.emx2.Column.column;
 import static org.molgenis.emx2.Row.row;
 import static org.molgenis.emx2.TableMetadata.table;
-import static org.molgenis.emx2.semantics.rdf.RDFTest.RDF_API_LOCATION;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
@@ -14,7 +14,6 @@ import org.junit.jupiter.api.Test;
 import org.molgenis.emx2.ColumnType;
 import org.molgenis.emx2.Database;
 import org.molgenis.emx2.Schema;
-import org.molgenis.emx2.semantics.RDFService;
 import org.molgenis.emx2.sql.TestDatabaseFactory;
 
 public class RDFWithCompositeKeysTest {
@@ -52,19 +51,44 @@ public class RDFWithCompositeKeysTest {
   }
 
   @Test
-  public void testRDFwithCompositeKeys() {
-    RDFService rdf = new RDFService("http://localhost:8080" + RDF_API_LOCATION);
+  void testRDFwithCompositeKeys() {
+    RDFService rdf = new RDFService("http://localhost:8080" + RDFTest.RDF_API_LOCATION);
     OutputStream outputStream = new ByteArrayOutputStream();
     rdf.describeAsRDF(
-        outputStream, RDF_API_LOCATION, null, null, null, List.of(schema).toArray(new Schema[1]));
+        outputStream,
+        RDFTest.RDF_API_LOCATION,
+        null,
+        null,
+        null,
+        List.of(schema).toArray(new Schema[1]));
     String result = outputStream.toString();
     assertTrue(result.contains(RDFWithCompositeKeysTest.class.getSimpleName()));
-
     String rowId =
+        "http://localhost:8080/RDFWithCompositeKeysTest/api/rdf/Samples/patient.firstName=Donald&patient.lastName=Duck&id=sample1";
+    String rowId2 =
         "http://localhost:8080/RDFWithCompositeKeysTest/api/rdf/Samples/patient.firstName=Donald&patient.lastName=Duck&id=sample2";
+    String check_composite_key_element =
+        "  <http://localhost:8080/RDFWithCompositeKeysTest/api/rdf/Patients/column/firstName>\n"
+            + "    \"Donald\";";
+    String check_composite_ref =
+        "  <http://localhost:8080/RDFWithCompositeKeysTest/api/rdf/Samples/column/patient> <http://localhost:8080/RDFWithCompositeKeysTest/api/rdf/Patients/patient.firstName=Donald&patient.lastName=Duck>;";
     assertTrue(result.contains(rowId));
+    assertTrue(result.contains(rowId2));
+    assertTrue(result.contains(check_composite_key_element));
+    assertTrue(result.contains(check_composite_ref));
 
     // should it also be possible to resolve these URIs?
     // but first rewrite the code because very confusing with passing of request/response
+    outputStream = new ByteArrayOutputStream();
+    rdf.describeAsRDF(
+        outputStream,
+        RDFTest.RDF_API_LOCATION,
+        schema.getTable("Samples"),
+        "patient.firstName=Donald&patient.lastName=Duck&id=sample2",
+        null,
+        List.of(schema).toArray(new Schema[1]));
+    result = outputStream.toString();
+    assertFalse(result.contains(rowId));
+    assertTrue(result.contains(rowId2));
   }
 }
