+++ backend/molgenis-emx2-rdf/src/test/resources/OntologyTableSemanticsTestFile.csv
